refactor: narrow config import surface
