+++ repowire/daemon/session_control.py
@@ -11,7 +11,7 @@ from typing import Any
 from fastapi import HTTPException
 
 from repowire.agent_backends import AgentResumePlan
-from repowire.config.models import AgentType
+from repowire.agent_types import AgentType
 from repowire.daemon.peer_registry import PeerRegistry
 from repowire.daemon.resume_safety import resolve_resume_safety
 from repowire.daemon.spawn_service import SpawnAttemptError, SpawnService
